Contracts scripts typings + JS-SDK cleanup
* bundle .d.ts with scripts for ethereum-contracts npm package
* add tsconfig.script.json file for scripts folder build
* fix up JS-SDK typings files
* fix JS-SDK typo
* add index.js in scripts as entrypoint for typings (test PR package build)
