+++ packages/ethereum-contracts/.gitignore
@@ -6,3 +6,4 @@
 /cache
 /docs/api
 /typechain-types
+/types
