+++ packages/ethereum-contracts/scripts/deploy-test-framework.js
@@ -1,0 +1,64 @@
+const {ethers} = require("hardhat");
+
+const SlotsBitmapLibraryArtifact = require("@superfluid-finance/ethereum-contracts/artifacts/contracts/libs/SlotsBitmapLibrary.sol/SlotsBitmapLibrary.json");
+const SuperfluidFrameworkDeployerArtifact = require("@superfluid-finance/ethereum-contracts/artifacts/contracts/utils/SuperfluidFrameworkDeployer.sol/SuperfluidFrameworkDeployer.json");
+
+const ERC1820Registry = require("./artifacts/ERC1820Registry.json");
+
+const ERC1820_ADDRESS = "0x1820a4b7618bde71dce8cdc73aab6c95905fad24";
+const ERC1820_BIN = ERC1820Registry.bin;
+const ERC1820_PAYLOAD =
+    "0xf90a388085174876e800830c35008080b909e5" +
+    ERC1820_BIN +
+    "1ba01820182018201820182018201820182018201820182018201820182018201820a01820182018201820182018201820182018201820182018201820182018201820";
+
+export async function deployERC1820(provider) {
+    console.log("Deploying ERC1820...");
+    const code = await provider.send("eth_getCode", [
+        ERC1820_ADDRESS,
+        "latest",
+    ]);
+    if (code === "0x") {
+        await provider.send("eth_sendRawTransaction", [ERC1820_PAYLOAD]);
+
+        console.log("ERC1820 registry successfully deployed");
+    }
+}
+
+const _getFactoryAndReturnDeployedContract = async (
+    contractName,
+    artifact,
+    signerOrOptions,
+    ...args
+) => {
+    console.log(`Deploying ${contractName}...`);
+    const ContractFactory = await ethers.getContractFactoryFromArtifact(
+        artifact,
+        signerOrOptions
+    );
+    const contract = await ContractFactory.deploy(...args);
+    await contract.deployed();
+    console.log(`${contractName} Deployed At:`, contract.address);
+    return contract;
+};
+
+export const deployTestFramework = async () => {
+    const signer = (await ethers.getSigners())[0];
+    await deployERC1820(ethers.provider);
+    const SlotsBitmapLibrary = await _getFactoryAndReturnDeployedContract(
+        "SlotsBitmapLibrary",
+        SlotsBitmapLibraryArtifact,
+        signer
+    );
+    const frameworkDeployer = await _getFactoryAndReturnDeployedContract(
+        "SuperfluidFrameworkDeployer",
+        SuperfluidFrameworkDeployerArtifact,
+        {
+            signer,
+            libraries: {
+                SlotsBitmapLibrary: SlotsBitmapLibrary.address,
+            },
+        }
+    );
+    return frameworkDeployer;
+};
