+++ packages/ethereum-contracts/scripts/index.js
@@ -1,0 +1,7 @@
+const deployAuxContracts = require("./deploy-aux-contracts");
+const deployTestFramework = require("./deploy-test-framework");
+
+module.exports = {
+    deployAuxContracts,
+    deployTestFramework
+};
