+++ packages/ethereum-contracts/tsconfig.json
@@ -101,6 +101,5 @@
         // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
         "skipLibCheck": true /* Skip type checking all .d.ts files. */
     },
-    "files": ["hardhat.config.ts"],
     "include": ["testsuites", "test", "./typechain-types", "scripts"]
 }
