+++ packages/ethereum-contracts/tsconfig.scripts.json
@@ -1,0 +1,24 @@
+{
+    // Change this to match your project
+    "include": ["scripts/*.js"],
+    "compilerOptions": {
+      // Tells TypeScript to read JS files, as
+      // normally they are ignored as source files
+      "allowJs": true,
+      // Generate d.ts files
+      "declaration": true,
+      // This compiler run should
+      // only output d.ts files
+      "emitDeclarationOnly": true,
+      // Types should go into this directory.
+      // Removing this would place the .d.ts files
+      // next to the .js files
+      "outDir": "types",
+      // go to js file when using IDE functions like
+      // "Go to Definition" in VSCode
+      "declarationMap": true,
+      // Emit additional JavaScript to ease support for importing CommonJS modules. 
+      // This enables 'allowSyntheticDefaultImports' for type compatibility.
+      "esModuleInterop": true
+    }
+  }
