+++ packages/js-sdk/src/User.d.ts
@@ -1,8 +1,8 @@
 import type { Framework } from './Framework';
 import type BN from 'bn.js';
-import type { Transaction } from 'web3';
-import type { Flow } from './ConstantFlowAgreementV1Helper';
-import type { Subscription } from './InstantDistributionAgreementV1Helper';
+import type { Transaction } from 'web3-core';
+import type { FlowInfo, FlowList } from './ConstantFlowAgreementV1Helper';
+import type { SubscriptionData } from './InstantDistributionAgreementV1Helper';
 
 // params options types
 export interface UserOptions {
@@ -41,7 +41,7 @@ export interface UserDetails {
         netFlow: string // numeric string
     }
     ida: {
-        subscriptions: Array<Subscription>;
+        subscriptions: Array<SubscriptionData>;
     }
 }
